adding an automatic error correction
correcting a dumb mistake and adding auto-correction for future errors.
When the chatroom was initialized in a channel and said channel got deleted its id remained in "chatroom_channels". This led to errors as the bot  was now trying to send messages in a deleted channel.
Fixed by checking the variable before trying to send messages, and deleting the entry in the file if the channel isn't valid.